add calls in container y add props in page
