+++ src/components/property/property-single-style/common/PropertyHeader.js
@@ -1,25 +1,8 @@
 "use client";
 
 //import listings from "@/data/listings";
-import { getListings } from "@/transform/propertiesTransform";
-import { useEffect, useState } from "react";
 
-const PropertyHeader = ({ id }) => {
-
-  const [listings, setListings] = useState([]);
-
-  useEffect(() => {
-    const loadListings = async () => {
-      try {
-        const items = await getListings();
-        setListings(items);
-      } catch (error) {
-        console.error("Error al cargar los items:", error);
-      }
-    };
-
-    loadListings();
-  }, []);
+const PropertyHeader = ({ id, listings }) => {
 
   const data = listings.filter((elm) => elm.id == id)[0] || listings[0];
 
@@ -46,12 +29,12 @@ const PropertyHeader = ({ id }) => {
               className="ff-heading bdrr1 fz15 pr10 ml10 ml0-sm bdrrn-sm"
               href="#">
               <i className="far fa-clock pe-2" />
-              {Number(new Date().getFullYear()) - Number(data.yearBuilding)} años
+              Construido el {data.yearBuilding}
             </a>
-            <a className="ff-heading ml10 ml0-sm fz15" href="#">
+            {/* <a className="ff-heading ml10 ml0-sm fz15" href="#">
               <i className="flaticon-fullscreen pe-2 align-text-top" />
               8721
-            </a>
+            </a> */}
           </div>
           <div className="property-meta d-flex align-items-center">
             <a className="text fz15" href="#">
@@ -74,7 +57,7 @@ const PropertyHeader = ({ id }) => {
       <div className="col-lg-4">
         <div className="single-property-content">
           <div className="property-action text-lg-end">
-            <div className="d-flex mb20 mb10-md align-items-center justify-content-lg-end">
+            {/* <div className="d-flex mb20 mb10-md align-items-center justify-content-lg-end">
               <a className="icon mr10" href="#">
                 <span className="flaticon-like" />
               </a>
@@ -87,7 +70,7 @@ const PropertyHeader = ({ id }) => {
               <a className="icon" href="#">
                 <span className="flaticon-printer" />
               </a>
-            </div>
+            </div> */}
             <h3 className="price mb-0">{data.price}</h3>
           </div>
         </div>
